// src/app/interfaces/SeasonRace.ts
export interface SeasonRace {
    seasonRaceId: number;
    seasonId: number;
    name: string;
    flagRace: string;
    laps: number;
    imageCircuit: string;
    firstPosition?: number;
    secondPosition?: number;
    thirdPosition?: number;
    fourthPosition?: number;
    fifthPosition?: number;
    sixthPosition?: number;
    seventhPosition?: number;
    eighthPosition?: number;
    ninthPosition?: number;
    tenthPosition?: number;
}